Payment Working

+++ app/DataTables/PaymentsDataTable.php
@@ -3,13 +3,13 @@
 namespace App\DataTables;
 
 use App\Models\Payment;
-use Illuminate\Database\Eloquent\Builder as QueryBuilder;
-use Illuminate\Support\Str;
-use Yajra\DataTables\EloquentDataTable;
-use Yajra\DataTables\Html\Builder as HtmlBuilder;
 use Yajra\DataTables\Html\Button;
 use Yajra\DataTables\Html\Column;
+use Yajra\DataTables\EloquentDataTable;
 use Yajra\DataTables\Services\DataTable;
+use Yajra\DataTables\Html\Builder as HtmlBuilder;
+use Illuminate\Database\Eloquent\Builder as QueryBuilder;
+use Barryvdh\DomPDF\Facade\Pdf;
 
 class PaymentsDataTable extends DataTable
 {
@@ -19,35 +19,23 @@ class PaymentsDataTable extends DataTable
      * @param QueryBuilder $query Results from query() method.
      * @return \Yajra\DataTables\EloquentDataTable
      */
-    public function dataTable(QueryBuilder $query): EloquentDataTable
+    public function dataTable(QueryBuilder $query)
     {
         $columns = array_column($this->getColumns(), 'data');
         return (new EloquentDataTable($query))
-            ->editColumn('check', function ($bookingPayment) {
-                return $bookingPayment;
+
+            ->editColumn('check', function ($payment) {
+                return $payment;
             })
-            ->editColumn('payment_debit', function ($bookingPayment) {
-                return ($bookingPayment->payment_debit) > 0 ? '$ ' . number_format($bookingPayment->payment_debit, 2) : '-';
+            ->editColumn('created_at', function ($payment) {
+                return editDateColumn($payment->created_at);
             })
-            ->editColumn('payment_credit', function ($bookingPayment) {
-                return ($bookingPayment->payment_credit) > 0 ? '$ ' . number_format($bookingPayment->payment_debit, 2) : '-';
+            ->editColumn('updated_at', function ($payment) {
+                return editDateColumn($payment->updated_at);
             })
-            ->editColumn('type', function ($bookingPayment) {
-                return Str::of($bookingPayment->type)->ucfirst();
-            })
-            ->editColumn('comments', function ($bookingPayment) {
-                $comments = Str::of($bookingPayment->comments);
-                return $comments->length() > 0 ? Str::of($bookingPayment->comments)->words(10) : '-';
-            })
-            ->editColumn('created_at', function ($bookingPayment) {
-                return editDateColumn($bookingPayment->created_at);
-            })
-            ->editColumn('updated_at', function ($bookingPayment) {
-                return editDateColumn($bookingPayment->updated_at);
-            })
-            ->editColumn('actions', function ($bookingPayment) {
-                return view('bookings.payments.actions', ['id' => $bookingPayment->id]);
-            })
+            // ->editColumn('actions', function ($payment) {
+            //     return view('bookings.actions', ['id' => $payment->id]);
+            // })
             ->setRowId('id')
             ->rawColumns(array_merge($columns, ['action', 'check']));
     }
@@ -58,56 +46,68 @@ class PaymentsDataTable extends DataTable
      * @param \App\Models\Payment $model
      * @return \Illuminate\Database\Eloquent\Builder
      */
-    public function query(Payment $model): QueryBuilder
+    public function query(Payment $model)
     {
-        return $model->newQuery();
+        return $model->newQuery()->where('booking_id', $this->booking_id);
     }
 
-    /**
-     * Optional method if you want to use html builder.
-     *
-     * @return \Yajra\DataTables\Html\Builder
-     */
     public function html(): HtmlBuilder
     {
+        $buttons = [];
+
+        if (auth()->user()->can('bookings.create')) {
+            $buttons[] = Button::raw('add-new')
+                ->addClass('btn btn-primary waves-effect waves-float waves-light m-1')
+                ->text('<i class="fa-solid fa-plus"></i>&nbsp;&nbsp;Add New')
+                ->attr([
+                    'onclick' => 'addNew()',
+                ]);
+        }
+
+        if (auth()->user()->can('bookings.export')) {
+            $buttons[] = Button::make('export')
+                ->addClass('btn btn-primary waves-effect waves-float waves-light dropdown-toggle m-1')
+                ->buttons([
+                    Button::make('print')->addClass('dropdown-item')->text('<i class="fa-solid fa-print"></i>&nbsp;&nbsp;Print'),
+                    Button::make('copy')->addClass('dropdown-item')->text('<i class="fa-solid fa-copy"></i>&nbsp;&nbsp;Copy'),
+                    Button::make('csv')->addClass('dropdown-item')->text('<i class="fa-solid fa-file-csv"></i>&nbsp;&nbsp;CSV'),
+                    Button::make('excel')->addClass('dropdown-item')->text('<i class="fa-solid fa-file-excel"></i>&nbsp;&nbsp;Excel'),
+                    Button::make('pdf')->addClass('dropdown-item')->text('<i class="fa-solid fa-file-pdf"></i>&nbsp;&nbsp;PDF'),
+                ]);
+        }
+
+        $buttons = array_merge($buttons, [
+            Button::make('reset')->addClass('btn btn-danger waves-effect waves-float waves-light m-1'),
+            Button::make('reload')->addClass('btn btn-primary waves-effect waves-float waves-light m-1'),
+        ]);
+
+        if (auth()->user()->can('bookings.payments.destroy')) {
+            $buttons[] = Button::raw('delete-selected')
+                ->addClass('btn btn-danger waves-effect waves-float waves-light m-1')
+                ->text('<i class="fa-solid fa-minus"></i>&nbsp;&nbsp;Delete Selected')
+                ->attr([
+                    'onclick' => 'deleteSelected()',
+                ]);
+        }
+
         return $this->builder()
-            ->addTableClass(['table-hover'])
-            ->setTableId('booking-payments-table')
+            ->setTableId('booking-table')
+            ->addTableClass('table-borderless table-striped table-hover')
             ->columns($this->getColumns())
+            ->minifiedAjax()
+            ->serverSide()
+            ->processing()
             ->deferRender()
-            ->scrollX()
             ->dom('BlfrtipC')
+            ->scrollX()
             ->pagingType('full_numbers')
             ->lengthMenu([
                 [30, 50, 70, 100, 120, 150, -1],
                 [30, 50, 70, 100, 120, 150, "All"],
             ])
             ->dom('<"card-header pt-0"<"head-label"><"dt-action-buttons text-end"B>><"d-flex justify-content-between align-items-center mx-0 row"<"col-sm-12 col-md-6"l><"col-sm-12 col-md-6"f>>t<"d-flex justify-content-between mx-0 row"<"col-sm-12 col-md-6"i><"col-sm-12 col-md-6"p>> C<"clear">')
-            ->buttons(
-                Button::raw('add-new')
-                    ->addClass('btn btn-relief-outline-primary waves-effect waves-float waves-light')
-                    ->text('<i class="bi bi-plus"></i> Add New')
-                    ->attr([
-                        'onclick' => 'addBookingPayment("' . encryptParams($this->booking_id) . '")',
-                    ]),
-                Button::make('export')->addClass('btn btn-relief-outline-secondary waves-effect waves-float waves-light dropdown-toggle')->buttons([
-                    Button::make('print')->addClass('dropdown-item'),
-                    Button::make('copy')->addClass('dropdown-item'),
-                    Button::make('csv')->addClass('dropdown-item'),
-                    Button::make('excel')->addClass('dropdown-item'),
-                    Button::make('pdf')->addClass('dropdown-item'),
-                ]),
-                Button::make('reset')->addClass('btn btn-relief-outline-danger waves-effect waves-float waves-light'),
-                Button::make('reload')->addClass('btn btn-relief-outline-primary waves-effect waves-float waves-light'),
-                Button::raw('delete-selected')
-                    ->addClass('btn btn-relief-outline-danger waves-effect waves-float waves-light')
-                    ->text('<i class="bi bi-trash3-fill"></i> Delete Selected')
-                    ->attr([
-                        'onclick' => 'deleteSelected()',
-                    ]),
-
-            )
-            // ->rowGroupDataSrc('type_id')
+            ->buttons($buttons)
+            // ->rowGroupDataSrc('parent_id')
             ->columnDefs([
                 [
                     'targets' => 0,
@@ -115,18 +115,18 @@ class PaymentsDataTable extends DataTable
                     'width' => '10%',
                     'orderable' => false,
                     'searchable' => false,
-                    'responsivePriority' => 0,
+                    'responsivePriority' => 3,
                     'render' => "function (data, type, full, setting) {
-                    var tableRow = JSON.parse(data);
-                    return '<div class=\"form-check\"> <input class=\"form-check-input dt-checkboxes\" onchange=\"changeTableRowColor(this)\" type=\"checkbox\" value=\"' + tableRow.id + '\" name=\"chkTableRow[]\" id=\"chkTableRow_' + tableRow.id + '\" /><label class=\"form-check-label\" for=\"chkTableRow_' + tableRow.id + '\"></label></div>';
-                }",
+                        var role = JSON.parse(data);
+                        return '<div class=\"form-check\"> <input class=\"form-check-input dt-checkboxes\" onchange=\"changeTableRowColor(this)\" type=\"checkbox\" value=\"' + role.id + '\" name=\"checkForDelete[]\" id=\"checkForDelete_' + role.id + '\" /><label class=\"form-check-label\" for=\"chkRole_' + role.id + '\"></label></div>';
+                    }",
                     'checkboxes' => [
                         'selectAllRender' =>  '<div class="form-check"> <input class="form-check-input" onchange="changeAllTableRowColor()" type="checkbox" value="" id="checkboxSelectAll" /><label class="form-check-label" for="checkboxSelectAll"></label></div>',
                     ]
                 ],
             ])
             ->orders([
-                [2, 'desc'],
+                [3, 'asc'],
             ]);
     }
 
@@ -137,15 +137,19 @@ class PaymentsDataTable extends DataTable
      */
     protected function getColumns(): array
     {
-        return [
-            Column::computed('check')->exportable(false)->printable(false)->width(60),
-            Column::make('payment_debit')->title('Debit'),
-            Column::make('payment_credit')->title('Credit'),
-            Column::make('type')->title('Type'),
-            Column::make('comments')->title('Comments'),
-            Column::make('created_at'),
-            Column::computed('actions')->exportable(false)->printable(false)->width(60)->addClass('text-center'),
+        $checkColumn = Column::computed('check')->exportable(false)->printable(false)->width(60)->addClass('text-nowarp');
+
+        if (auth()->user()->can('bookings.payments.destroy')) {
+            $checkColumn->addClass('disabled');
+        }
+
+        $columns = [
+            $checkColumn,
+            Column::make('created_at')->addClass('text-nowarp'),
+            Column::make('updated_at')->addClass('text-nowarp'),
+            Column::computed('actions')->exportable(false)->printable(false)->width(60)->addClass('text-center text-nowrap'),
         ];
+        return $columns;
     }
 
     /**
@@ -155,6 +159,17 @@ class PaymentsDataTable extends DataTable
      */
     protected function filename(): string
     {
-        return 'Booking_Payments_' . date('YmdHis');
+        return 'booking_' . date('YmdHis');
+    }
+
+    /**
+     * Export PDF using DOMPDF
+     * @return mixed
+     */
+    public function pdf()
+    {
+        $data = $this->getDataForPrint();
+        $pdf = Pdf::loadView($this->printPreview, ['data' => $data])->setOption(['defaultFont' => 'sans-serif']);
+        return $pdf->download($this->filename() . '.pdf');
     }
 }
